Move form builder to own module and separate files
