+++ src/modules/formBuilder/components/formTree/types.ts
@@ -1,0 +1,9 @@
+import { SvgIconComponent } from '@mui/icons-material';
+
+export type FormItemDisplay = {
+  icon: SvgIconComponent;
+  text: string;
+  backgroundColor: string;
+  textColor: string;
+  hoverColor: string;
+};
